Build break fix
Summary:
(1) Integer size correction (mac build break)
(2) snprint usage in Windows (windows build break)

Test Plan: Build in windows and mac

Reviewers: sdong

Subscribers: andrewkr, dhruba

Differential Revision: https://reviews.facebook.net/D60927

+++ utilities/persistent_cache/block_cache_tier_file.cc
@@ -38,7 +38,7 @@ Status NewRandomAccessCacheFile(Env* const env, const std::string& filepath,
 //
 // BlockCacheFile
 //
-Status BlockCacheFile::Delete(size_t* size) {
+Status BlockCacheFile::Delete(uint64_t* size) {
   Status status = env_->GetFileSize(Path(), size);
   if (!status.ok()) {
     return status;
